# backend/openapi.yaml
openapi: 3.0.3
info:
  title: MedLine API
  description: >
    Middleware de gestión de farmacia hospitalaria: RBAC, inventario con FEFO
    y Stock Útil, libro de movimientos, Motor de Compra Inteligente, gestión
    de proveedores, devoluciones/logística inversa, historia clínica
    (farmacia) con versionado, y auditoría inmutable.


    Además de los endpoints REST documentados aquí, el servidor expone un
    canal **WebSocket** en `ws://<host>/ws?token=<JWT>` para notificaciones en
    tiempo real (nuevos movimientos, cambios de estado de órdenes de compra,
    devoluciones procesadas/rechazadas). OpenAPI no modela WebSockets; se deja
    esta nota como referencia. Cada evento llega sólo a los sockets cuyo
    usuario tiene el permiso del recurso correspondiente.
  version: "1.0.0"
servers:
  - url: /api
    description: API relativa (mismo origen que el frontend en desarrollo/producción)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Token emitido por POST /auth/login. Embebe el set de permisos del
        usuario; el backend lo revalida en cada endpoint protegido con
        requirePermission(...).
  schemas:
...
    StockReturn:
      type: object
      properties:
        id: { type: string }
        code: { type: string, example: 'RET-2026-000123' }
        quantity: { type: integer }
        reason: { type: string, enum: [EXPIRED, DAMAGED, UNUSED, WRONG_DISPENSE, RECALL, OTHER] }
        disposition: { type: string, nullable: true, enum: [RESTOCK, DISCARD, RETURN_TO_SUPPLIER] }
        status: { type: string, enum: [PENDING, PROCESSED, REJECTED] }
        sourceLocation: { type: string, nullable: true }
        note: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        processedAt: { type: string, format: date-time, nullable: true }
        item: { type: object, properties: { id: { type: string }, name: { type: string } } }
        batch: { type: object, nullable: true, properties: { id: { type: string }, batchNumber: { type: string } } }
        requestedBy: { type: object, nullable: true, properties: { id: { type: string }, fullName: { type: string } } }
        processedBy: { type: object, nullable: true, properties: { id: { type: string }, fullName: { type: string } } }
...
  parameters:
    ItemIdFilter:
      name: itemId
      in: query
      schema: { type: string }
      description: Filtra por un ítem específico
...
security:
  - bearerAuth: []
paths:
...
  /returns:
    get:
      tags: [Returns]
      summary: Listar devoluciones, filtro opcional por estado/ítem (requiere returns:read)
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [PENDING, PROCESSED, REJECTED] }
        - $ref: '#/components/parameters/ItemIdFilter'
      responses:
        '200':
          description: Devoluciones
          content:
            application/json:
              schema:
                type: object
                properties:
                  returns: { type: array, items: { $ref: '#/components/schemas/StockReturn' } }
    post:
      tags: [Returns]
      summary: Registrar una devolución en estado PENDING (requiere returns:create)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [itemId, quantity, reason]
              properties:
                itemId: { type: string }
                batchId: { type: string }
                quantity: { type: integer, minimum: 1 }
                reason: { type: string, enum: [EXPIRED, DAMAGED, UNUSED, WRONG_DISPENSE, RECALL, OTHER] }
                sourceLocation: { type: string }
                note: { type: string }
      responses:
        '201': { description: Devolución registrada }
  /returns/{id}:
    get:
      tags: [Returns]
      summary: Detalle de una devolución (requiere returns:read)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200': { description: Devolución }
        '404': { description: No encontrada }
  /returns/{id}/process:
    post:
      tags: [Returns]
      summary: >
        Decide la disposición final (requiere returns:process). RESTOCK
        reingresa a stock (crea/actualiza lote + movimiento RETURN) en una
        transacción; requiere batchNumber/expirationDate/physicalLocation.
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [disposition]
              properties:
                disposition: { type: string, enum: [RESTOCK, DISCARD, RETURN_TO_SUPPLIER] }
                batchNumber: { type: string }
                expirationDate: { type: string, format: date }
                physicalLocation: { type: string }
                note: { type: string }
      responses:
        '200': { description: Devolución procesada }
        '409': { description: 'Ya fue procesada o rechazada' }
  /returns/{id}/reject:
    post:
      tags: [Returns]
      summary: Rechazar una devolución pendiente (requiere returns:process)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200': { description: Devolución rechazada }
        '409': { description: 'Ya fue procesada o rechazada' }
...